carving out common functions to new file
